Support configurable connection timeout
This adds support for a configurable connection timeout that defaults to
3 seconds. I also changed the return value in the init function under
error conditions to `{:stop, reason}` from `{:error, reason}`. I think
this is correct because `{:error, reason}` generated:

```
** (EXIT from #PID<0.208.0>) bad return value: {:error, :timeout}
```

I could also not find support for `{:error, _}` in the docs:
https://hexdocs.pm/elixir/GenServer.html#c:init/1

This was tricky to test because calling `Gnat.start_link` killed the
test immediately. Hopefully calling `Gnat.init` instead is a viable way
to test the behavior.