docs: add hex badge and link to hexdocs
Currently a user has to look up hexdocs by going to hexdoc and looking for gnat. The repo is called nats.ex so this might be a little misleading.

This commit adds both a badge displaying the current version (with a link to docs) and an explicit box to show the link to hexdocs.